refactor: 배틀페이지(index)에서 battle:closed 이벤트 수신 처리 로직 useBattleProgress훅으로 이동
